requirements.txt
python-chess==1.999
torch>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
flask>=3.0.0
